# src/main/java/com/example/demo/model/User.java
package com.example.demo.model;

import io.swagger.v3.oas.annotations.media.Schema;
import jakarta.persistence.Entity;
import jakarta.persistence.GeneratedValue;
import jakarta.persistence.GenerationType;
import jakarta.persistence.Id;
import jakarta.persistence.Table;
import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

@Entity
@Table(name="user")
@Schema(description = "Details about a user")
public class User {
	@Id
    @GeneratedValue(strategy = GenerationType.IDENTITY)
	@Schema(description = "The unique ID of the user", example = "1")
	private Long id;
	
	@Schema(description = "The user's name", example = "John Doe")
	private String name;
	
	@Schema(description = "The user's email address", example = "john.doe@example.com")
    private String email;
	public Long getId() {
		return id;
	}
	public void setId(Long id) {
		this.id = id;
	}
	public String getName() {
		return name;
	}
	public void setName(String name) {
		this.name = name;
	}
	public String getEmail() {
		return email;
	}
	public void setEmail(String email) {
		this.email = email;
	}
    public User(Long id, String name, String email) {
		super();
		this.id = id;
		this.name = name;
		this.email = email;
	}
    public User() {
	}
   

}